cakeshop/requirements.txt
amqp==2.5.2
asgiref==3.2.3
autopep8==1.5
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.2
defusedxml==0.6.0
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
drf-yasg==1.17.1
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
pyparsing==2.4.6
Pyphen==0.9.5
pytz==2019.3
PyYAML==5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
zipp==3.1.0
